requirements.txt
a2wsgi==1.10.10
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosmtplib==5.0.0
aiosqlite==0.22.0
alembic==1.17.2
annotated-types==0.7.0
anyio==4.12.0
apache-airflow-providers-amazon==9.18.1
apache-airflow-providers-common-compat==1.10.1
apache-airflow-providers-common-io==1.7.0
apache-airflow-providers-common-sql==1.30.1
apache-airflow-providers-google==12.7.0
apache-airflow-providers-http==5.6.1
apache-airflow-providers-postgres==6.5.1
apache-airflow-providers-smtp==2.4.1
apache-airflow-providers-standard==1.10.1
argcomplete==3.6.3
asgiref==3.11.0
asn1crypto==1.5.1
asyncpg==0.31.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
boto3==1.42.0
botocore==1.41.6
cadwyn==5.4.5
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorlog==6.10.1
contourpy==1.3.3
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.3
cycler==0.12.1
dask==2025.12.0
Deprecated==1.3.1
dill==0.4.0
dnspython==2.8.0
duckdb==1.4.2
email-validator==2.3.0
fastapi==0.117.1
fastapi-cli==0.0.16
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
geopandas==1.0.1
googleapis-common-protos==1.72.0
greenback==1.2.1
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonpath-ng==1.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy-object-proxy==1.12.0
libcst==1.8.6
linkify-it-py==2.0.3
locket==1.0.0
lockfile==0.12.2
lxml==5.4.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.1.1
matplotlib==3.10.8
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
msgspec==0.20.0
multidict==6.7.0
natsort==8.4.0
numpy==2.3.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
partd==1.4.2
pathspec==0.12.1
pendulum==3.1.0
pillow==12.0.0
pluggy==1.6.0
ply==3.11
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
psycopg2-binary==2.9.11
pyarrow==22.0.0
PyAthena==3.22.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydynamodb==0.8.0
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.10.1
pyogrio==0.12.1
pyparsing==3.2.5
pyproj==3.7.2
pyspainmobility==1.0.4
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-argparse==1.7.2
rich-toolkit==0.17.0
rpds-py==0.30.0
s3transfer==0.16.0
sagemaker_studio==1.0.23
scramp==1.4.6
setproctitle==1.3.7
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.45
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.1
sqlean.py==3.50.4.5
sqlparse==0.5.4
starlette==0.48.0
structlog==25.5.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.2.0
text-unidecode==1.3
toolz==1.1.0
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.6.2
uuid6==2025.0.1
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
watchtower==3.4.0
websockets==15.0.1
wirerope==1.0.0
wrapt==2.0.1
yarl==1.22.0
zipp==3.23.0
